requirements.txt
sentence_transformers==2.2.2
langchain
flask
pypdf
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_community
langchain-google-genai
langchain_experimental
langchain_google_vertexai
-e .
